Move template util methods to utils/templates Rename utils/template and utils/validation to utils/templates and utils/validations
